+++ game/game-scene/src/main/java/com/whk/scene/server/SceneServerManager.java
@@ -27,14 +27,14 @@ public class SceneServerManager extends ServerManager {
 
     private GameDateConfig gameDateConfig;
 
-    private final Set<Integer> newAddGateServerIds = new ConcurrentHashSet<>();
+    private final Set<Long> newAddGateServerIds = new ConcurrentHashSet<>();
 
     public void init(GameDateConfig gameDateConfig, DiscoveryClient discoveryClient) {
         this.discoveryClient = discoveryClient;
         this.gameDateConfig = gameDateConfig;
     }
 
-    public void gateNoticeUpdateServer(int gateServerId) {
+    public void gateNoticeUpdateServer(long gateServerId) {
         newAddGateServerIds.add(gateServerId);
         updateGate();
         if (!getGroupServers(ServerType.GATE).keySet().containsAll(newAddGateServerIds)) {
